Move SQLStatementContextConverter to shardingsphere-infra-binder module (#7992)
* Update javadoc of SQLStatementContextConverter

* Move SQLStatementContextConverter to shardingsphere-infra-binder module

* Rename package name from convert to converter
